feat: WordSearchModule 생성 및 AppModule 에 적용
