+++ src/word/word.controller.ts
@@ -23,14 +23,6 @@ import {
 } from './dto/word-detail.dto';
 import { RequestWordListDto, ResponseWordListDto } from './dto/word-list.dto';
 import {
-	RequestWordRelatedSearchDto,
-	ResponseWordRelatedSearchDto,
-} from './dto/word-related-search.dto';
-import {
-	RequestWordSearchDto,
-	ResponseWordSearchDto,
-} from './dto/word-search.dto';
-import {
 	RequestWordUserLikeDto,
 	ResponseWordUserLikeDto,
 } from './dto/word-user-like.dto';
@@ -86,47 +78,6 @@ export class WordController {
 			userId: user.id,
 		});
 		return await this.wordService.getWordUserLike(wordUserLikeDto);
-	}
-
-	@ApiDocs({
-		summary: '특정 키워드와 연관된 단어 목록을 조회합니다.',
-		response: {
-			statusCode: HttpStatus.OK,
-			schema: ResponseWordRelatedSearchDto,
-			isPaginated: true,
-		},
-	})
-	@Get('/search/related')
-	async findByRelatedSearch(
-		@Query() requestWordRelatedSearchDto: RequestWordRelatedSearchDto,
-	) {
-		return await this.wordService.getWordByRelatedKeyword(
-			requestWordRelatedSearchDto,
-		);
-	}
-
-	@ApiDocs({
-		summary: '특정 키워드 검색에 대한 단어 목록을 조회합니다.',
-		response: {
-			statusCode: HttpStatus.OK,
-			schema: ResponseWordSearchDto,
-			isPaginated: true,
-		},
-	})
-	@UseInterceptors(UserInformationInterceptor)
-	@Get('/search')
-	async findBySearch(
-		@AuthenticatedUser() user: User,
-		@Query() requestWordSearchDto: RequestWordSearchDto,
-	) {
-		const wordSearchDto = plainToInstance(RequestWordSearchDto, {
-			userId: user?.id,
-			keyword: requestWordSearchDto.keyword,
-			page: requestWordSearchDto.page,
-			limit: requestWordSearchDto.limit,
-		});
-
-		return await this.wordService.getWordByKeyword(wordSearchDto);
 	}
 
 	@ApiDocs({
